coordinator/requirements.txt
requests==2.28.1
Pillow==9.4.0
streamlit==1.35.0